requirements.txt
Django==1.8.5
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.4.3
Markdown==2.6.6
